+++ pkg/util/resourcecontainer/resource_container_linux.go
@@ -43,7 +43,3 @@ func RunInResourceContainer(containerName string) error {
 
 	return manager.Apply(os.Getpid())
 }
-
-func ApplyRLimitForSelf(maxOpenFiles uint64) {
-	syscall.Setrlimit(syscall.RLIMIT_NOFILE, &syscall.Rlimit{Max: maxOpenFiles, Cur: maxOpenFiles})
-}
